requirements.txt
absl-py==0.10.0
aiogrpc==1.8
anki-vector==0.6.0
appdirs==1.4.4
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
attrs==20.3.0
autopy==4.0.0
bazel==0.0.0.20200723
black==20.8b1
boto3==1.17.48
botocore==1.20.48
cached-property==1.5.2
cachetools==4.1.1
camera==1.3.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.3
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
decorator==4.4.2
distlib==0.3.2
dlib==19.22.0
docutils==0.17.1
enum34==1.1.10
evdev==1.4.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
gast==0.2.2
gin-config==0.1.4
glcontext==2.3.4
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.31.0
h11==0.12.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
importlib-resources==5.1.4
imutils==0.5.4
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Kivy==2.0.0
Kivy-Garden==0.1.4
kivy-garden.mapview==1.0.5
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
mapview==1.0.5
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mediapipe==0.8.3
moderngl==5.6.4
MouseInfo==0.1.3
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.19.3
oauthlib==3.1.0
objloader==0.2.0
opencv-python==4.4.0.42
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.1
pathspec==0.8.0
pep517==0.6.0
Pillow==7.2.0
pkg-resources==0.0.0
progressbar2==3.52.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycairo==1.19.1
pycparser==2.20
pygame==2.0.1
PyGetWindow==0.0.9
pyglet==1.5.17
Pygments==2.9.0
pylint==2.6.0
PyMsgBox==1.0.7
pynput==1.7.3
PyOpenGL==3.1.5
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-for-android==2020.6.2
python-utils==2.4.0
python-xlib==0.29
python3-xlib==0.15
pytoml==0.1.21
PyTweening==1.0.3
pytz==2020.1
PyWavefront==1.3.3
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.6
scikit-image==0.17.2
scipy==1.4.1
sh==1.14.2
six==1.15.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tifffile==2020.9.3
toml==0.10.1
tqdm==4.60.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.13.4
virtualenv==20.4.7
vtk==9.0.1
Werkzeug==1.0.1
wrapt==1.12.1
wxPython==4.1.1
zipp==3.1.0
